Gemini_pdf_thinking_fast_slow/requirements.txt
chromadb>=0.4.18
google.generativeai
tqdm
python-dotenv
pdfquery
